test: `testkit.eq` exige deux arguments du même type
0023 a fonctionné : la section 3 de 0007_push.sql passe en entier, pour la
première fois. Le nouvel échec est de nouveau dans le test, pas dans le
produit.

  `function testkit.eq(bigint, integer, unknown) does not exist`

  `testkit.eq` est `eq(anyelement, anyelement, text)` : les deux valeurs
  comparées doivent être du MÊME type, faute de quoi PostgreSQL ne trouve
  aucune surcharge. `testkit.count()` renvoie un bigint, et `v_before` était
  déclarée `integer`.

  Le détail qui rend le défaut particularly insidieux : l'AFFECTATION accepte
  bigint → integer sans bruit. `v_before := testkit.count(…)` fonctionne, et
  le type ne se révèle qu'à la comparaison, sous la forme d'un message qui ne
  parle ni de l'assertion ni de ce qu'elle vérifie. `v_before` est désormais
  déclarée `bigint`, et le commentaire explique pourquoi.

check-sql-statique.py, cinquième contrôle
  Les deux valeurs de `testkit.eq` doivent être du même type. Il ne porte que
  sur `supabase/tests` : les migrations n'appellent pas testkit.

  Le contrôleur ne conclut que sur ce qu'il peut déterminer avec certitude. Un
  littéral s'accorde sur l'autre argument — `testkit.eq(testkit.count(…), 0)`
  est valide — donc il ne peut pas être la cause d'un conflit, et le silence
  vaut mieux qu'une devinette. Même principe que pour `returns table` : un
  contrôle qui signale un faux conflit se fait ignorer avant d'avoir trouvé un
  vrai défaut.

  Il a aussi attrapé une erreur de ma propre rédaction : mon nouveau motif
  s'appelait `RE_VARIABLE`, comme le filtre `p_`/`v_` qui écartait les faux
  positifs de `extract(dow from p_day)`. Le second l'avait écrasé, et trois
  fonctions parfaitement saines sont ressorties en défaut. Le contrôle
  fonctionne : il a refusé la version cassée du fichier, sur les deux lignes.

Vérifié : dépôt vert (76 fonctions effectives, 24 migrations, 9 tests).

+++ supabase/tests/0007_push.sql
@@ -430,7 +430,13 @@ declare
   v_alice uuid;
   v_consumed integer;
   v_report jsonb;
-  v_before integer;
+  -- `bigint`, et non `integer` : `testkit.count()` renvoie un bigint, et
+  -- `testkit.eq` est `eq(anyelement, anyelement, text)` — les deux arguments
+  -- doivent être du MÊME type, sinon PostgreSQL ne trouve aucune surcharge et
+  -- répond « function testkit.eq(bigint, integer, unknown) does not exist ».
+  -- L'affectation, elle, accepte bigint → integer sans bruit : le défaut
+  -- n'apparaissait qu'à la comparaison.
+  v_before bigint;
 begin
   select user_id into v_alice from testkit.fx where key = 'camille';
